+++ lib/pages/suggest_remedies/suggest_remedies_ui.dart
