+++ lib/screens/bank_details/bank_detail_binding.dart
