+++ lib/screens/bank_details/bank_details_ui.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
 
-import '../../../common/TextFieldCustom.dart';
+import '../../../common/text_field_custom.dart';
 import '../../../common/app_textstyle.dart';
 import '../../../common/strings.dart';
 import '../../../gen/assets.gen.dart';
