+++ lib/screens/blocked_user/blocked_user_bindings.dart
@@ -1,6 +1,6 @@
 import 'package:get/get.dart';
 
-import 'BlockUserController.dart';
+import 'block_user_controller.dart';
 
 class BlockedUserBinding extends Bindings {
   @override
