+++ lib/screens/dashboard/dashboard_bindings.dart
@@ -1,7 +1,7 @@
-import 'package:divine_astrologer/screens/pages/home/home_controller.dart';
 import 'package:get/get.dart';
 
-import '../wallet/wallet_controller.dart';
+import '../../pages/home/home_controller.dart';
+import '../../pages/wallet/wallet_controller.dart';
 import 'dashboard_controller.dart';
 
 class DashboardBinding extends Bindings {
