+++ lib/screens/home_screen_options/refer_astrologer/refer_astrologer_ui.dart
@@ -7,8 +7,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
 
-import '../../../common/TextFieldCustom.dart';
-import '../../../common/customLightYellowButton.dart';
+import '../../../common/text_field_custom.dart';
+import '../../../common/custom_light_yellow_btn.dart';
 
 class ReferAnAstrologer extends GetView<ReferAstrologerController> {
   const ReferAnAstrologer({Key? key}) : super(key: key);
