+++ lib/screens/price_change/price_change_binding.dart
