+++ lib/screens/price_history/price_history_binding.dart
